refactor: move routes to routes.ts for reload
